% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_script.R
\name{current_script}
\alias{current_script}
\title{Gets the pathname of the currently running startup script}
\usage{
current_script()
}
\value{
A character string
}
\description{
Gets the pathname of the currently running startup script
}
