% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\alias{uninstall}
\title{Install and uninstall support for .Renviron.d and .Rprofile.d startup
directories}
\usage{
install(
  file = rprofile_user(),
  backup = TRUE,
  overwrite = FALSE,
  path = dirname(file),
  make_dirs = TRUE,
  quiet = FALSE
)

uninstall(file = rprofile_user(), backup = TRUE, quiet = FALSE)
}
\arguments{
\item{file}{The pathname where to create or update the \file{.Rprofile}
file.}

\item{backup}{If \code{TRUE}, a timestamped backup copy of the original file is
created before modifying / overwriting it, otherwise not.  If the backup
fails, then an error is produced and the R startup file is unmodified.}

\item{overwrite}{If the R startup file already exists, then \code{FALSE} (default)
appends the startup code to the end of the file. If \code{TRUE}, any pre-existing
R startup file is overwritten.}

\item{path}{The folder in which to create the \file{.Renviron.d} and
\file{.Rprofile.d} directory.}

\item{make_dirs}{If \code{TRUE} (default), directories \file{.Renviron.d/} and
\file{.Rprofile.d/} are created in folder \code{path}, if missing.}

\item{quiet}{If \code{FALSE} (default), detailed messages are generated,
otherwise not.}
}
\value{
The pathname of the R startup file modified.
}
\description{
Install and uninstall support for \file{.Renviron.d} and \file{.Rprofile.d}
startup directories by appending or removing a line of code to the
\file{~/.Rprofile} file.
}
\section{Functions}{
\itemize{
\item \code{install()}: injects a \code{tryCatch(startup::startup(), ...)} call to
the \file{.Rprofile} file, which is created if missing.

\item \code{uninstall()}: Remove calls to \code{startup::startup()} and similar.

}}
