% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ess.R
\name{is_ess}
\alias{is_ess}
\title{Checks whether running R via Emacs Spekas Statistics (ESS) or not}
\usage{
is_ess()
}
\value{
A logical
}
\description{
Checks whether running R via Emacs Spekas Statistics (ESS) or not
}

