% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_radian.R
\name{is_radian}
\alias{is_radian}
\title{Check if running R via radian (formerly known as rtichoke and rice)}
\usage{
is_radian()
}
\value{
A logical
}
\description{
Check if running R via radian (formerly known as rtichoke and rice)
}
\references{
\enumerate{
\item radian - A 21st-century R console (previously known as rtichoke and rice),
\url{https://github.com/randy3k/radian}
}
}
