% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_rice.R
\name{is_rice}
\alias{is_rice}
\title{Checks whether running R via Rice or not}
\usage{
is_rice()
}
\value{
A logical
}
\description{
Checks whether running R via Rice or not
}
\references{
\enumerate{
\item Rice - A Command-Line Interface for R,
\url{https://github.com/randy3k/Rice}
}
}
