% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_rscript.R
\name{is_rscript}
\alias{is_rscript}
\title{Checks if running R via Rscript or not}
\usage{
is_rscript()
}
\value{
A logical
}
\description{
\emph{WARNING: This function does not work - it always returns FALSE [1].}
}
\details{
This functions use \code{basename(commandArgs()[1])} to infer whether or not
\R was launched via \file{Rscript}.  On Windows, both \file{Rscript} and
\file{Rscript.exe} are recognized and the check is case insensitive.
}
\references{
\enumerate{
\item \url{https://github.com/HenrikBengtsson/startup/issues/97}
}
}
