% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_rstudio.R
\name{is_rstudio}
\alias{is_rstudio}
\title{Checks whether running R via RStudio or not}
\usage{
is_rstudio()
}
\value{
A logical
}
\description{
Checks whether running R via RStudio or not
}
\references{
\itemize{
\item Kevin Ushey (RStudio), Check if R is running in RStudio,
Stackoverflow, 2016-09-23,
\url{http://stackoverflow.com/questions/12389158/check-if-r-is-running-in-rstudio#comment66636507_35849779} #nolint
}
}
