% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_rstudio.R
\name{is_rstudio_console}
\alias{is_rstudio_console}
\title{Checks if running R in RStudio Console or RStudio Terminal}
\usage{
is_rstudio_console()
}
\value{
A logical
}
\description{
Checks if running R in RStudio Console or RStudio Terminal
}
\references{
\enumerate{
\item Kevin Ushey (RStudio), Check if R is running in RStudio,
Stackoverflow, 2016-09-23,
\url{https://stackoverflow.com/questions/12389158/check-if-r-is-running-in-rstudio#comment66636507_35849779}
\item Jonathan McPherson (RStudio), Programmatically detect RStudio Terminal vs RStudio Console?, RStudio Community - RStudio IDE, 2018-01-10, \url{https://community.rstudio.com/t/programmatically-detect-rstudio-terminal-vs-rstudio-console/4107}
}
}
