% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_rtichoke.R
\name{is_rtichoke}
\alias{is_rtichoke}
\title{Checks whether running R via rtichoke (formerly rice) or not}
\usage{
is_rtichoke()
}
\value{
A logical
}
\description{
Checks whether running R via rtichoke (formerly rice) or not
}
\references{
\enumerate{
\item rtichoke - A 21 century R console (previously known as rice),
\url{https://github.com/randy3k/rtichoke}
}
}
