% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_vscode.R
\name{is_vscode}
\alias{is_vscode}
\title{Checks if running is R via Visual Studio Code (VSCode)}
\usage{
is_vscode()
}
\value{
A logical
}
\description{
Checks if running is R via Visual Studio Code (VSCode)
}
