% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_webr.R
\name{is_webr}
\alias{is_webr}
\title{Checks if running R via webR}
\usage{
is_webr()
}
\value{
A logical
}
\description{
Checks if running R via webR
}
\references{
\enumerate{
\item WebR - R in the Browser, \url{https://docs.r-wasm.org/webr/latest/}
}
}
