% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_wine.R
\name{is_wine}
\alias{is_wine}
\title{Check if running R on Windows using Wine}
\usage{
is_wine()
}
\value{
A logical
}
\description{
Check if running R on Windows using Wine
}
\references{
\enumerate{
\item Wine Developer FAQ, How can I detect Wine?,
\url{https://gitlab.winehq.org/wine/wine/-/wikis/Developer-FAQ#How_can_I_detect_Wine.3F}
\item Jeff Zaroyko, Detecting Wine, Wine Devel mailing list, 2008-09-29,
\url{https://www.winehq.org/pipermail/wine-devel/2008-September/069389.html}
}
}
