% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytoolbox.R
\name{mytoolbox}
\alias{mytoolbox}
\title{Organize Your Toolbox}
\usage{
mytoolbox(expr = ls(all.names = TRUE), name = "startup::mytoolbox",
  pos = NULL)
}
\arguments{
\item{expr}{An \R expression to be evaluated inside the 'mytoolbox'
environment.  The default is to list all tools in the toolbox.}

\item{name}{(optional) The name of the toolbox environment.}

\item{pos}{(optional) A numeric. If specified, the toolbox will be moved
to this location on the search path.}
}
\description{
Organize Your Toolbox
}
\details{
One or more of your tools in your toolbox may be masked by functions in
\emph{attached} packages.  To avoid this, move your toolbox to the front of
the \link[base:search]{search} path by calling \code{startup::mytoolbox(pos = 2L)}.
}
\examples{
## Add tools to the toolbox
startup::mytoolbox({
  ## Quit R without questions
  Q <- function(save = "no", ...) quit(save = save, ...)

  ## List _all_ variables
  ll <- startup::partial(ls, all.names = TRUE)
})

## Add another tool to the toolbox
startup::mytoolbox({
  ## Install package in current directory from source
  install <- startup::partial(utils::install.packages,
                              pkgs = ".", repos = NULL)
})

## List all tools in the toolbox
startup::mytoolbox()

}
