% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_session_exit.R
\name{on_session_exit}
\alias{on_session_exit}
\title{Register R expressions and functions to be evaluated when R terminates}
\usage{
on_session_exit(fcn = NULL, append = TRUE, replace = FALSE)
}
\arguments{
\item{fcn}{An R function or an R expression.  If an expression, it will
automatically we wrapped up in an anonymous function.}

\item{append}{If TRUE (default), the function will be evaluated after
previously registered ones, otherwise prepended.}

\item{replace}{if TRUE, the function replaces any previously registered
ones, otherwise it will be added (default).}
}
\value{
(invisible) the list of registered functions.
}
\description{
Register R expressions and functions to be evaluated when R terminates
}
\details{
This function registers one or more functions to be called when the R
session terminates.  All of them are evaluated in a local environment.
These functions are evaluated without exception handlers, which means that
if one produces an error, then none of the following will be evaluated.

To list currently registered functions, call \code{fcns <- on_session_exit()}.
To remove all registered functions, call \code{on_session_exit(replace = TRUE)}.

The function works by recording all \code{fcn}:s in an internal list which will
be evaluated via a custom function that is called when the global
environment is garbage collected, which happens at the very end of the R
shutdown process.
Contrary to \code{.Last()} and \code{.Last.sys()}, which may not be called if
\code{quit(runLast = FALSE)} is used, functions registered via
\code{on_session_exit()} are always processed.
}
