% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_session_start.R
\name{on_session_start}
\alias{on_session_start}
\title{Register an R expression to be evaluated at the end of the R startup process}
\usage{
on_session_start(expr, substitute = TRUE, append = TRUE, replace = FALSE)
}
\arguments{
\item{expr, substitute}{An R expression. If \code{substitute = TRUE}, \code{expr}
is automatically substituted.}

\item{append}{If TRUE, the expression is added to the end of the list
of expression to be evaluated, otherwise prepended.}

\item{replace}{if TRUE, the expression replaces all existing expressions
previously added, otherwise it will be added (default).}
}
\value{
(invisible) the list of start functions
}
\description{
Register an R expression to be evaluated at the end of the R startup process
}
\details{
This function registers one or more R expressions to be evaluated at the
very end of the R startup process.  All expressions are evaluated in a
local environment.  The expressions are evaluated without exception
handlers, which means that if one expression produces an error, then
none of the following expression will be evaluated.

The function works by recording all expressions \code{expr} in an internal list
which will be evaluated via a custom \code{.First()} function created in the
global environment. Any other \code{.First()} function on the search path,
including a pre-existing \code{.First()} function in the global environment,
is called at the end after registered expressions have been called.
}
