% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial.R
\name{partial}
\alias{partial}
\title{Partially Apply a Function by Setting One Or More of the Arguments}
\usage{
partial(fun, ...)
}
\arguments{
\item{fun}{A function.}

\item{\ldots}{Named arguments of \code{fun} to be set or modified.}
}
\value{
A function.
}
\description{
Partially Apply a Function by Setting One Or More of the Arguments
}
\examples{
## Create ll() as ls() but with all.names = TRUE
ll <- startup::partial(ls, all.names = TRUE)

}
