% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agenda.R
\name{reset_when_cache}
\alias{reset_when_cache}
\title{Reset all or parts of the "when" cache}
\usage{
reset_when_cache(when = c("once", "hourly", "daily", "weekly",
  "fortnightly", "monthly"))
}
\arguments{
\item{when}{(character vector) Specifies for which "when" frequencies
the cache should be reset.  The default is to reset all of them.}
}
\value{
(invisible) The pathnames of the "when" cache files removed.
}
\description{
Reset all or parts of the "when" cache
}
