% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup_session_info.R
\name{startup_session_info}
\alias{startup_session_info}
\title{Record R session information as options}
\usage{
startup_session_info(reuse = TRUE, record = TRUE)
}
\arguments{
\item{reuse}{If \code{TRUE} and any of the \code{"startup.session.*"}
options are already recorded, then those are used, otherwise their values
are inferred from the information currently available.}

\item{record}{If \code{TRUE}, the information gathered is recorded as \R
options with prefix \code{"startup.session."} accessible as regular options,
e.g. \code{getOption("startup.session.startupdir")}.}
}
\value{
Returns invisibly a named list of the options prefixed
\code{"startup.session."} including:
\describe{
  \item{\code{startup.session.startdir}}{(character) the directory where
    the \R session was launched from}
  \item{\code{startup.session.id}}{(character) a unique ID for the current
    \R session}
  \item{\code{startup.session.dumpto}}{(character) a value that can be used
    for argument \code{dumpto} of
    \code{\link[utils:dump.frames]{dump.frames()}}}
    (also for dumping to file)
}
}
\description{
Record R session information as options
}
\examples{
info <- startup::startup_session_info()
info

}
