% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{toolbox}
\alias{toolbox}
\title{Organize Your Own Toolboxes}
\usage{
toolbox(expr = NULL, name = "default", pos = NULL)
}
\arguments{
\item{expr}{An \R expression to be evaluated inside an 'toolbox'
environment.}

\item{name}{(optional) The name of the toolbox environment.}

\item{pos}{(optional) A numeric. If specified, the toolbox will be moved
to this location on the search path.  If \code{pos = 0}, the toolbox is removed.}
}
\value{
The value of the evaluated expression.  If \code{expr} is \code{NULL}
(default), then a named list of the content of all toolboxes.
}
\description{
Organize Your Own Toolboxes
}
\details{
One or more of your tools in your toolbox may be masked by functions in
\emph{attached} packages.  To avoid this, move your toolbox to the front of
the \link[base:search]{search} path by calling \code{startup::toolbox(pos = 2L)}.

An empty toolbox environment is removed automatically.
To remove a toolbox, either empty it or move it to position zero or less.
}
\examples{
## Add tools to the toolbox
startup::toolbox({
  ## Quit R without questions
  Q <- function(save = "no", ...) quit(save = save, ...)

  ## List _all_ variables
  ll <- startup::partial(ls, all.names = TRUE)
})

## Add another tool to the toolbox
startup::toolbox({
  ## Install package in current directory from source
  install <- startup::partial(utils::install.packages,
                              pkgs = ".", repos = NULL)
})

## List all tools in all toolboxes
startup::toolbox()

## Remove one of the tools
startup::toolbox({
  rm(list = "Q")
})

\dontrun{
## Remove toolbox by moving it to position zero
startup::toolbox(pos = 0)
}

}
