% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_files.R
\name{find_rprofile}
\alias{find_renviron}
\alias{find_rprofile}
\title{Locates the .Rprofile and .Renviron files used during the startup of R}
\usage{
find_rprofile(all = FALSE)

find_renviron(all = FALSE)
}
\description{
Locates the .Rprofile and .Renviron files used during the startup of R
}
\section{Functions}{
\itemize{
\item \code{find_rprofile}: Locates the \file{.Rprofile} file used during \R startup.

\item \code{find_renviron}: Locates the \file{.Renviron} file used during \R startup.
}}
\keyword{internal}

