% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\alias{uninstall}
\title{Install and uninstall startup to .Rprofile}
\usage{
install(path = "~", backup = TRUE, debug = FALSE)

uninstall(path = "~", backup = TRUE, debug = FALSE)
}
\arguments{
\item{path}{The path where to create / update the \file{.Rprofile} file.}

\item{backup}{If \code{TRUE}, a timestamped backup copy of the original file is created before modifying it, otherwise not.}

\item{debug}{If \code{TRUE}, debug messages are outputted, otherwise not.}
}
\description{
Install and uninstall startup to .Rprofile
}
\section{Functions}{
\itemize{
\item \code{install}: Appends a \code{startup::startup()} call to the \file{.Rprofile}.

\item \code{uninstall}: Remove calls to \code{startup::startup()} and similar.
}}

