% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pqr.R
\name{is_pqr}
\alias{is_pqr}
\title{Checks if running pqR (A Pretty Quick Version of R)}
\usage{
is_pqr()
}
\value{
A logical
}
\description{
Checks if running pqR (A Pretty Quick Version of R)
}
\references{
\enumerate{
\item pqR - a pretty quick version of R,
\url{http://www.pqr-project.org/}
\item GitHub repository for 'pqR'
\url{https://github.com/radfordneal/pqR}
}
}
