% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_radian.R
\name{is_radian}
\alias{is_radian}
\title{Checks whether running R via radian (formerly known as rtichoke and rice) or not}
\usage{
is_radian()
}
\value{
A logical
}
\description{
Checks whether running R via radian (formerly known as rtichoke and rice) or not
}
\references{
\enumerate{
\item radian - A 21 century R console (previously known as rtichoke and rice),
\url{https://github.com/randy3k/radian}
}
}
