% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_wine.R
\name{is_wine}
\alias{is_wine}
\title{Checks whether running R on Windows using Wine or not}
\usage{
is_wine()
}
\value{
A logical
}
\description{
Checks whether running R on Windows using Wine or not
}
\references{
\itemize{
 \item Wine Developer FAQ, How can I detect Wine?,
  \url{https://wiki.winehq.org/Developer_FAQ#How_can_I_detect_Wine.3F}
 \item Jeff Zaroyko, Detecting Wine, Wine Devel mailing list, 2008-09-29,
  \url{https://www.winehq.org/pipermail/wine-devel/2008-September/069387.html}
}
}
