% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytoolbox.R
\name{mytoolbox}
\alias{mytoolbox}
\title{Organize Your Toolbox}
\usage{
mytoolbox(expr = ls(all.names = TRUE))
}
\arguments{
\item{expr}{An \R expression to be evaluated inside the 'mytoolbox'
environment.  The default is to list all tools in the toolbox.}
}
\description{
Organize Your Toolbox
}
\examples{
## Add tools to the toolbox
startup::mytoolbox({
  ## Quit R without questions
  Q <- function(save = "no", ...) quit(save = save, ...)

  ## List _all_ variables
  ll <- startup::partial(ls, all.names = TRUE)
})

## Add another tool to the toolbox
startup::mytoolbox({
  ## Install package in current directory from source
  install <- startup::partial(utils::install.packages,
                              pkgs = ".", repos = NULL)
})

## List all tools in the toolbox
startup::mytoolbox()

}
