% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytoolbox.R
\name{mytoolbox}
\alias{mytoolbox}
\title{Organize Your Toolbox}
\usage{
mytoolbox(expr = ls(all.names = TRUE), name = "startup::mytoolbox")
}
\arguments{
\item{expr}{An \R expression to be evaluated inside the 'mytoolbox'
environment.  The default is to list all tools in the toolbox.}

\item{name}{(optional) The name of the toolbox environment.}
}
\description{
Organize Your Toolbox
}
\details{
Your tools will be kept first on the \link[base:search]{search} path.
For instance, if one of your tools have the same name as a function in
an \emph{attached} package, your tool will be used.
This is achieved by always moving the 'mytoolbox' environment (where your
tools sits) to the front of the \link[base:search]{search} path whenever the
\code{startup::mytoolbox()} function is called.
}
\examples{
## Add tools to the toolbox
startup::mytoolbox({
  ## Quit R without questions
  Q <- function(save = "no", ...) quit(save = save, ...)

  ## List _all_ variables
  ll <- startup::partial(ls, all.names = TRUE)
})

## Add another tool to the toolbox
startup::mytoolbox({
  ## Install package in current directory from source
  install <- startup::partial(utils::install.packages,
                              pkgs = ".", repos = NULL)
})

## List all tools in the toolbox (and move toolbox to the front)
startup::mytoolbox()

}
