% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_session_enter.R
\name{on_session_enter}
\alias{on_session_enter}
\title{Register R expressions and functions to be evaluated at the end of the R startup process}
\usage{
on_session_enter(
  expr = NULL,
  substitute = TRUE,
  append = TRUE,
  replace = FALSE
)
}
\arguments{
\item{expr, substitute}{An R expression or a function.
If \code{substitute = TRUE}, \code{expr} is automatically substituted.
\code{NULL} expressions are ignored.}

\item{append}{If TRUE (default), the expression or the function will be
evaluated after previously registered entries, otherwise prepended.}

\item{replace}{if TRUE, the expression or the function replaces all
existing ones previously added, otherwise it will be added (default).}
}
\value{
(invisible) the list of registered expressions and functions.
}
\description{
Register R expressions and functions to be evaluated at the end of the R startup process
}
\details{
This function registers one or more R expressions or functions to be
evaluated at the very end of the R startup process.  All of them are
evaluated in a local environment.  These expressions and functions are
evaluated without exception handlers, which means that if one produces an
error, then none of the following will be evaluated.

To list currently registered expressions and functions, call
\code{exprs <- on_session_enter()}. To remove all registered entries, call
\code{on_session_enter(replace = TRUE)}.

The function works by recording all \code{expr} in an internal list which will
be evaluated via a custom \code{\link[base:.First]{.First()}} function
created in the global environment. Any other \code{.First()} function on the
search path, including a pre-existing \code{.First()} function in the global
environment, is called at the end after registered expressions have been
called.
}
