% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restart.R
\name{restart}
\alias{restart}
\title{Restarts R}
\usage{
restart(status = 0L, rcmd = NULL, args = NULL, envvars = NULL,
  as = NULL, debug = NA)
}
\arguments{
\item{status}{An integer specifying the exit code of the current
\R session.}

\item{rcmd}{A character string specifying the command for launching \R.
The default is the same as used to launch the current \R session, i.e.
\code{\link[base:commandArgs]{commandArgs()[1]}}.}

\item{args}{A character vector specifying zero or more command-line
arguments to be passed to the system call of \code{rcmd}.
If \code{NULL}, then \code{\link[base:commandArgs]{commandArgs()[-1]}} is used.
To not pass any arguments, use \code{args = character(0L)}.}

\item{envvars}{A named character vector of environment variables to
be set when calling \R.  Alternatively, if not named the strings must
be of format \code{var=value}.}

\item{as}{A character string specify a predefined set of \code{rcmd}, \code{args},
and \code{envvars}.  If \code{as = "R CMD build"}, then the \code{R CMD build} environment
is emulated as far as possible.}

\item{debug}{If \code{TRUE}, debug messages are outputted, otherwise not.}
}
\description{
Restarts R by quitting the current \R session and launching a new one.
}
\examples{
\dontrun{
  startup::restart(envvars = c(R_STARTUP_DEBUG = TRUE))

  ## Mimic R CMD build
  startup::restart(as = "R CMD build")

  ## ... which is short for
  startup::restart(args = c("--no-restore"),
                   envvars = c(R_DEFAULT_PACKAGES="", LC_COLLATE="C"))
}

}
