% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staRtup.R
\name{staRtup}
\alias{staRtup}
\title{StaRtup - create the base folders you need for your R project quickly}
\usage{
staRtup()
}
\value{

}
\description{
StaRtup - create the base folders you need for your R project quickly
}
\examples{
staRtup()
}
