% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{startup}
\alias{renviron}
\alias{rprofile}
\alias{startup}
\title{Simplified Initialization at Start of an R Session}
\usage{
startup(paths = c(".", "~"), unload = TRUE, debug = NA)

renviron(paths = c(".", "~"), unload = FALSE, debug = NA)

rprofile(paths = c(".", "~"), unload = FALSE, debug = NA)
}
\arguments{
\item{paths}{Character vector of paths where to locate the \file{.Renviron.d/} and \file{.Rprofile.d/} directories.}

\item{unload}{If \code{TRUE}, then the package is unloaded afterward, otherwise not.}

\item{debug}{If \code{TRUE}, debug messages are outputted, otherwise not.}
}
\description{
Initiates R using all files under \file{.Renviron.d/}
and / or \file{.Rprofile.d/} directories (or in subdirectories)
thereof.  This is done in addition the \file{.Renviron} and
\file{.Rprofile} files supported by the
\link[base:Startup]{default R startup process}.
Directories \file{.Renviron.d/} and \file{.Rprofile.d/} may
be located in the current directory and / or the user's home
directory.
}
\section{Functions}{
\itemize{
\item \code{startup}: \code{renviron()} followed by \code{rprofile()} and then the package is unloaded

\item \code{renviron}: Initiate using \file{.Renviron.d/} files

\item \code{rprofile}: Initiate using \file{.Rprofile.d/} files
}}
\examples{
\dontrun{
# The most common way to use the package
startup::startup()

# Initiate any ./.Renviron.d/ and ~/.Renviron.d/ files
startup::renviron()

# Initiate only ~/.Rprofile.d/ files
startup::rprofile(paths = "~")

# Initiate .Renviron.d/ files then .Rprofile.d/ files
startup::renviron()$rprofile()
}

}

