% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renviron.R, R/rprofile.R, R/startup.R
\name{renviron}
\alias{renviron}
\alias{rprofile}
\alias{startup}
\title{Simplified Initialization at Start of an R Session}
\usage{
renviron(paths = c("~", "."), unload = FALSE, debug = NA)

rprofile(paths = c("~", "."), unload = FALSE, debug = NA)

startup(paths = c("~", "."), unload = TRUE, debug = NA)
}
\arguments{
\item{paths}{Character vector of paths where to locate the \file{.Renviron.d/} and \file{.Rprofile.d/} directories.}

\item{unload}{If \code{TRUE}, then the package is unloaded afterward, otherwise not.}

\item{debug}{If \code{TRUE}, debug messages are outputted, otherwise not.}
}
\description{
Initiates R using all files under \file{.Renviron.d/}
and / or \file{.Rprofile.d/} directories
(or in subdirectories thereof).
Any \file{.Renviron.d/} and \file{.Rprofile.d/} directories
in user's home directory (`~`) are first processed followed
by any in the in the current directory (`.`).
}
\details{
The above is done in addition the \file{.Renviron} and
\file{.Rprofile} files that re supported by the built-in
\link[base:Startup]{startup process} of \R.
}
\section{Functions}{
\itemize{
\item \code{renviron}: Initiate using \file{.Renviron.d/} files

\item \code{rprofile}: Initiate using \file{.Rprofile.d/} files

\item \code{startup}: \code{renviron()} followed by \code{rprofile()} and then the package is unloaded
}}
\examples{
\dontrun{
# The most common way to use the package is to add
# the following call to the ~/.Rprofile file.
startup::startup()

# For finer control of on exactly what files are used
# functions renviron() and rprofile() are also available:

# Initiate any ./.Renviron.d/ and ~/.Renviron.d/ files
startup::renviron()

# Initiate only ~/.Rprofile.d/ files
startup::rprofile(paths = "~")

# Initiate .Renviron.d/ files then .Rprofile.d/ files
startup::renviron()$rprofile()
}

}

