% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{startup}
\alias{everything}
\alias{renviron}
\alias{rprofile}
\alias{startup}
\title{Simplified Initialization at Start of an R Session}
\usage{
everything()

renviron()

rprofile()
}
\description{
Simplified Initialization at Start of an R Session

Initiate .Renviron.d/ files then .Rprofile.d/ files

Initiate .Renviron.d/ files

Initiate .Rprofile.d/ files
}
\examples{
\dontrun{
# Initiate just .Renviron.d/ files
startup::renviron()

# Initiate just .Rprofile.d/ files
startup::rprofiled()

# Initiate .Renviron.d/ files then .Rprofile.d/ files
startup::renviron()$rprofile()

# Initiate everything (as above in the same order)
startup::everything()
  
# Initiate everything and unload package afterward
startup::everything()$unload()
}

}

