% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{warn}
\alias{warn}
\title{Produce a warning}
\usage{
warn(..., call. = FALSE, immediate. = TRUE, domain = NULL)
}
\arguments{
\item{..., domain}{The message and optionally the domain used for
translation.  The \ldots arguments are passed to \link[base:sprintf]{base::sprintf} to
create the message string.}

\item{call.}{If \link[base:TRUE]{base::TRUE}, the call is included in the warning
message, otherwise not.}

\item{immediate.}{If \link[base:TRUE]{base::TRUE}, the warning is output immediately,
otherwise not.}
}
\value{
Nothing.
}
\description{
Produce a warning
}
